+++ spec/src/mock/circular_child.ts
@@ -1,7 +1,7 @@
-import { Serialize } from '../@lib/serialazy';
+import { Serializable } from '../@lib/serialazy';
 
 import Parent from './circular_parent';
 
 export default class Child {
-    @Serialize({ optional: true }) public parent: Parent;
+    @Serializable.Prop({ optional: true }) public parent: Parent;
 }